+++ example/lib/src/biz/biz2/flutter10/flutter10.page.dart
@@ -20,29 +20,45 @@ class Flutter10Page extends NavigatorStatefulPage {
 }
 
 class _Flutter10PageState extends State<Flutter10Page>
-    with NavigatorPageLifecycleMixin, AutomaticKeepAliveClientMixin {
+    with AutomaticKeepAliveClientMixin {
   @override
   Widget build(final BuildContext context) {
     super.build(context);
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        title: const Text('thrio_build_example',
-            style: TextStyle(color: Colors.black)),
-        leading: context.showPopAwareWidget(const IconButton(
-          color: Colors.black,
-          tooltip: 'back',
-          icon: Icon(Icons.arrow_back_ios),
-          onPressed: ThrioNavigator.pop,
-        )),
-        systemOverlayStyle: SystemUiOverlayStyle.dark,
-      ),
-      body: const SizedBox(
-        height: 100,
-        child: Text('flutter 10'),
-      ),
-    );
+    return const _TestConstWidget();
   }
+
+  @override
+  bool get wantKeepAlive => true;
+}
+
+class _TestConstWidget extends StatefulWidget {
+  const _TestConstWidget();
+
+  @override
+  State<_TestConstWidget> createState() => __TestConstWidgetState();
+}
+
+class __TestConstWidgetState extends State<_TestConstWidget>
+    with NavigatorPageLifecycleMixin {
+  @override
+  Widget build(final BuildContext context) => Scaffold(
+        appBar: AppBar(
+          backgroundColor: Colors.white,
+          title: const Text('thrio_build_example',
+              style: TextStyle(color: Colors.black)),
+          leading: context.showPopAwareWidget(const IconButton(
+            color: Colors.black,
+            tooltip: 'back',
+            icon: Icon(Icons.arrow_back_ios),
+            onPressed: ThrioNavigator.pop,
+          )),
+          systemOverlayStyle: SystemUiOverlayStyle.dark,
+        ),
+        body: const SizedBox(
+          height: 100,
+          child: Text('flutter 10'),
+        ),
+      );
 
   @override
   void didAppear(final RouteSettings settings) {
@@ -53,7 +69,4 @@ class _Flutter10PageState extends State<Flutter10Page>
   void didDisappear(final RouteSettings settings) {
     ThrioLogger.d('flutter10 didDisappear: $settings ');
   }
-
-  @override
-  bool get wantKeepAlive => true;
 }