+++ example/lib/main.dart
@@ -1,40 +1,26 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_thrio/flutter_thrio.dart';
 
 import 'src/app.dart' as app;
 
 Future<void> main() async {
   ThrioLogger.v('main');
-  FlutterError.onError = (final details) async {
-    Zone.current.handleUncaughtError(
-      details.exception,
-      details.stack ?? StackTrace.empty,
-    );
-  };
-  runZoned<void>(app.main);
+  runZonedGuarded(app.main, (final error, final stack) {
+    Zone.current.handleUncaughtError(error, stack);
+  });
 }
 
 @pragma('vm:entry-point')
 Future<void> biz1() async {
-  FlutterError.onError = (final details) async {
-    Zone.current.handleUncaughtError(
-      details.exception,
-      details.stack ?? StackTrace.empty,
-    );
-  };
-  runZoned<void>(app.biz1);
+  runZonedGuarded(app.biz1, (final error, final stack) {
+    Zone.current.handleUncaughtError(error, stack);
+  });
 }
 
 @pragma('vm:entry-point')
 Future<void> biz2() async {
-  FlutterError.onError = (final details) async {
-    Zone.current.handleUncaughtError(
-      details.exception,
-      details.stack ?? StackTrace.empty,
-    );
-  };
-  runZoned<void>(app.biz2);
+  runZonedGuarded(app.biz2, (final error, final stack) {
+    Zone.current.handleUncaughtError(error, stack);
+  });
 }